Vendor prebuilt OpenSSL 1.0.2u binaries
Compile OpenSSL for iOS and macOS using a well-known script [1]. We are
using OpenSSL 1.0.2u -- the latest version of 1.0.2t branch which should
include a fix for CVE-2019-1547. (Unfortunately, OpenSSL 1.1.1 does not
quite work with Themis on Apple systems yet.)

We are opting for direct vendoring to provide a hot fix, this is not
sustainable in the long run. Unfortunately, we have to vendor because
OpenSSL dependency we use is not actively maintained and does not
contain recent updates in OpenSSL.